refactor(mcp): split layer ownership/visibility gating into layer-access.ts
layer-store.ts crossed the ~400-line module guideline after the
storage-model changes; the access-matrix helpers are a coherent module
of their own (storage vs gating).
